+++ Cyotek.Windows.Forms.ColorPicker/GimpPaletteSerializer.cs
@@ -16,6 +16,28 @@ namespace Cyotek.Windows.Forms
   /// </summary>
   public class GimpPaletteSerializer : PaletteSerializer
   {
+    #region Overridden Properties
+
+    /// <summary>
+    /// Gets the default extension for files generated with this palette format.
+    /// </summary>
+    /// <value>The default extension for files generated with this palette format.</value>
+    public override string DefaultExtension
+    {
+      get { return ".gpl"; }
+    }
+
+    /// <summary>
+    /// Gets a descriptive name of the palette format
+    /// </summary>
+    /// <value>The descriptive name of the palette format.</value>
+    public override string Name
+    {
+      get { return "GIMP Palette"; }
+    }
+
+    #endregion
+
     #region Overridden Members
 
     /// <summary>