+++ Cyotek.Windows.Forms.ColorPicker/Properties/AssemblyInfo.cs
@@ -13,7 +13,6 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: Guid("de546619-59b3-438e-8b5a-3d149e146b22")]
 [assembly: CLSCompliant(true)]
-[assembly: AssemblyVersion("1.0.1.1")]
-[assembly: AssemblyFileVersion("1.0.1.1")]
+[assembly: AssemblyVersion("1.0.2.3")]
+[assembly: AssemblyFileVersion("1.0.2.3")]
 
-// ## UPDATED